# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CATAMS API
  description: |
    Casual Academic Time Allocation Management System (CATAMS) API specification.
    
    This API provides endpoints for managing academic staff time allocation, including:
    - User authentication and management
    - Timesheet creation and management
    - Approval workflow management
    
    The system supports three main roles:
    - **LECTURER**: Can create and approve timesheets for tutors
    - **TUTOR**: Can view their timesheets and request modifications
    - **ADMIN**: Has full system access including HR functions
  version: 1.0.0
  contact:
    name: CATAMS Development Team
    email: dev@catams.edu.au
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.catams.edu.au
    description: Production server
...
paths:
...
  /approvals:
    post:
      tags:
        - Approvals
      summary: Perform approval action
      description: |
        Perform an approval action on a timesheet (approve, reject, or request modification).
        
        **Access rules**:
        - **LECTURER**: Can approve/reject timesheets for their courses
        - **TUTOR**: Can request modifications to their own timesheets
        - **ADMIN**: Can perform any approval action
        
        **Workflow states**:
        - DRAFT → PENDING_LECTURER_APPROVAL (submit for approval)
        - PENDING_LECTURER_APPROVAL → APPROVED or REJECTED or MODIFICATION_REQUESTED
        - MODIFICATION_REQUESTED → PENDING_LECTURER_APPROVAL (after tutor addresses)
        - APPROVED → PENDING_HR_REVIEW (auto-transition)
        - PENDING_HR_REVIEW → FINAL_APPROVED or REJECTED (ADMIN/HR only)
      operationId: performApprovalAction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalActionRequest'
            examples:
              approve_timesheet:
                summary: Approve timesheet
                value:
                  timesheetId: 789
                  action: "APPROVE"
                  comment: "Approved - hours and description look correct"
              reject_timesheet:
                summary: Reject timesheet
                value:
                  timesheetId: 789
                  action: "REJECT"
                  comment: "Rejected - hours seem excessive for the described work"
              request_modification:
                summary: Request modification
                value:
                  timesheetId: 789
                  action: "REQUEST_MODIFICATION"
                  comment: "Please provide more detailed description of the tutorial work"
      responses:
        '200':
          description: Approval action completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalActionResponse'
              examples:
                approval_success:
                  summary: Successful approval action
                  value:
                    timesheetId: 789
                    action: "APPROVE"
                    newStatus: "APPROVED"
                    approverId: 999
                    approverName: "Dr. John Smith"
                    comment: "Approved - hours and description look correct"
                    timestamp: "2025-08-01T16:30:00Z"
                    nextSteps:
                      - "Timesheet automatically submitted to HR for final review"
        '400':
          description: Invalid approval action or business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_transition:
                  summary: Invalid state transition
                  value:
                    timestamp: "2025-08-01T10:15:30.123Z"
                    status: 400
                    error: "INVALID_STATE_TRANSITION"
                    message: "Cannot approve timesheet in current state: DRAFT"
                    path: "/api/approvals"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_authorized:
                  summary: Not authorized for this timesheet
                  value:
                    timestamp: "2025-08-01T10:15:30.123Z"
                    status: 403
                    error: "ACCESS_DENIED"
                    message: "You are not authorized to approve this timesheet"
                    path: "/api/approvals"
        '404':
          description: Timesheet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    ApprovalStatus:
      type: string
      enum:
        - DRAFT
        - PENDING_LECTURER_APPROVAL
        - APPROVED
        - REJECTED
        - MODIFICATION_REQUESTED
        - PENDING_HR_REVIEW
        - FINAL_APPROVED
      description: |
        Approval status of a timesheet:
        - **DRAFT**: Initial state, being prepared
        - **PENDING_LECTURER_APPROVAL**: Submitted for lecturer review
        - **APPROVED**: Approved by lecturer, ready for HR
        - **REJECTED**: Rejected by lecturer or HR
        - **MODIFICATION_REQUESTED**: Lecturer requested changes
        - **PENDING_HR_REVIEW**: Awaiting final HR approval
        - **FINAL_APPROVED**: Final approval by HR, ready for payment
      example: "PENDING_LECTURER_APPROVAL"
    ApprovalActionRequest:
      type: object
      required:
        - timesheetId
        - action
      properties:
        timesheetId:
          type: integer
          format: int64
          description: ID of the timesheet to act upon
          example: 789
        action:
          $ref: '#/components/schemas/ApprovalAction'
        comment:
          type: string
          description: Optional comment explaining the action
          example: "Approved - hours and description look correct"
          maxLength: 500
    ApprovalAction:
      type: string
      enum:
        - SUBMIT_FOR_APPROVAL
        - APPROVE
        - REJECT
        - REQUEST_MODIFICATION
      description: |
        Type of approval action:
        - **SUBMIT_FOR_APPROVAL**: Submit draft timesheet for approval
        - **APPROVE**: Approve the timesheet
        - **REJECT**: Reject the timesheet
        - **REQUEST_MODIFICATION**: Request changes to the timesheet
      example: "APPROVE"
    ApprovalActionResponse:
      type: object
      required:
        - timesheetId
        - action
        - newStatus
        - approverId
        - approverName
        - timestamp
      properties:
        timesheetId:
          type: integer
          format: int64
          description: ID of the timesheet that was acted upon
          example: 789
        action:
          $ref: '#/components/schemas/ApprovalAction'
        newStatus:
          $ref: '#/components/schemas/ApprovalStatus'
        approverId:
          type: integer
          format: int64
          description: ID of the user who performed the action
          example: 999
        approverName:
          type: string
          description: Name of the user who performed the action
          example: "Dr. John Smith"
        comment:
          type: string
          description: Comment provided with the action
          example: "Approved - hours and description look correct"
        timestamp:
          type: string
          format: date-time
          description: When the action was performed
          example: "2025-08-01T16:30:00Z"
        nextSteps:
          type: array
          description: Information about what happens next in the workflow
          items:
            type: string
          example:
            - "Timesheet automatically submitted to HR for final review"
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
          example: "2025-08-01T10:15:30.123Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error code for programmatic handling
          example: "VALIDATION_FAILED"
        message:
          type: string
          description: Human-readable error message
          example: "Validation failed: email: Email format is invalid"
        path:
          type: string
          description: Request path that caused the error
          example: "/api/auth/login"

# Security configuration
security:
  - bearerAuth: []